tests/test_requirements.txt
flask>=2.0.0
pillow>=9.0.0  # Для заглушки изображений в локал сайте
bs4